pyproject.toml
[project]
name = "code-debugger-env"
version = "1.0.0"
description = "Python Code Debugger OpenEnv Environment — hackathon submission"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2",
    "openai",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
